v0.2.3 (#20)
* Improve styling of no recent migration actions messaging
* Add template name to recent actions list
* Move actionRunner function utils to their own file
* Added a `functions/README.md` to start function documentation
