--- functions/mockData/actionRunner/test.json
@@ -1,52 +1,0 @@
-{
-  "createdAt" : 1517128955690,
-  "createdBy" : "FScHm5UzLtfytd014vYBmNIFSTZ2",
-  "inputValues" : [
-    {
-      "databaseURL" : "https://fireadmin-33d82.firebaseio.com",
-      "environmentKey": "LGxjF9f9BXcBMyAHHIWH",
-      "fullPath" : "serviceAccounts/aZJDcgty1Ofske1kusTDEFsjWeh2/ZHg98Qsk8aTDuhqwYIjH/fireadmin-be0f35bca620.json"
-    },{
-    "databaseURL" : "https://fireadmin-stage.firebaseio.com",
-    "environmentKey": "GVoP6U8NZkG47DTrawgX",
-    "fullPath" : "serviceAccounts/aZJDcgty1Ofske1kusTDEFsjWeh2/ZHg98Qsk8aTDuhqwYIjH/fireadmin-stage-e1ce9e573bd1.json"
-  },
-  "archived"],
-  "projectId" : "ZHg98Qsk8aTDuhqwYIjH",
-  "startedAt" : 1517128966709,
-  "status" : "started",
-  "template" : {
-    "inputs" : [ {
-      "description" : "Creds for first RTDB",
-      "name" : "Service Account 1",
-      "type" : "serviceAccount",
-      "variableName" : "serviceAccount1"
-    }, {
-      "description" : "Creds for second RTDB instance",
-      "name" : "Service Account 2",
-      "type" : "serviceAccount",
-      "variableName" : "serviceAccount2"
-    },
-    {
-      "description" : "Path to copy",
-      "name" : "Copy Path",
-      "type" : "userInput",
-      "variableName" : "copyPath"
-    } ],
-    "steps" : [ {
-      "description" : "Copy users from one RTDB instance to another under archived collection",
-      "dest" : {
-        "path" : 2,
-        "pathType": "input",
-        "resource" : "rtdb"
-      },
-      "name" : "Archive Users From RTDB",
-      "src" : {
-        "path" : 2,
-        "pathType": "input",
-        "resource" : "rtdb"
-      },
-      "type" : "copy"
-    } ]
-  }
-}
