--- functions/mockData/mockEnvironment.json
@@ -1,7 +1,0 @@
-{
-  "name": "testing",
-  "projectId": "nq3efmWhto512agiNOgI",
-  "serviceAccount": {
-    "fullPath": "serviceAccounts/FScHm5UzLtfytd014vYBmNIFSTZ2/nq3efmWhto512agiNOgI/serviceAccount.json"
-  }
-}
