+++ functions/src/callGoogleApi/constants.js
@@ -1,0 +1,6 @@
+export const eventPathName = 'googleApi'
+
+export const SCOPES = [
+  'https://www.googleapis.com/auth/devstorage.full_control',
+  'https://www.googleapis.com/auth/cloud-platform'
+]
