+++ functions/src/indexActionTemplates/index.js
@@ -1,5 +1,5 @@
+import * as functions from 'firebase-functions'
 import { createIndexFunc } from '../utils/search'
-const functions = require('firebase-functions')
 
 // Updates the search index for action templates when template is created or updated.
 export default functions.firestore
