+++ functions/src/storageFileToRTDB/index.js
@@ -1,11 +1,11 @@
+import * as functions from 'firebase-functions'
 import fs from 'fs-extra'
 import os from 'os'
 import path from 'path'
 import * as admin from 'firebase-admin'
 import mkdirp from 'mkdirp-promise'
-const functions = require('firebase-functions')
+
 const gcs = require('@google-cloud/storage')()
-const bucket = gcs.bucket(functions.config().firebase.storageBucket)
 
 const eventPathName = 'fileToDb'
 
@@ -22,17 +22,24 @@ async function copyFileToRTDB(event) {
   const eventData = event.data.val()
   const { filePath, databasePath, keepPushKey = false } = eventData
   const tempLocalFile = path.join(os.tmpdir(), filePath)
+  // Create temp directory
   const tempLocalDir = path.dirname(tempLocalFile)
   await mkdirp(tempLocalDir)
-  // Create Temporary directory and download file to that folder
+
+  const bucket = gcs.bucket(functions.config().firebase.storageBucket)
+
+  // Download file to temporary directory
   await bucket.file(filePath).download({ destination: tempLocalFile })
+
   // Read the file
   const fileData = await fs.readJson(filePath)
   console.log('File data loaded, writing to database', event.data.val())
+
   // Write File data to DB
   await event.data.adminRef.ref.root
     .child(`${databasePath}/${keepPushKey ? event.params.pushId : ''}`)
     .set(fileData)
+
   // Mark request as complete
   await event.data.adminRef.ref.root
     .child(`responses/${eventPathName}/${event.params.pushId}`)
