+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fireadmin",
-  "version": "0.2.2",
+  "version": "0.2.3",
   "description": "Application for Managing Firebase Applications. Includes support for multiple environments and data migrations.",
   "main": "index.js",
   "scripts": {
