+++ src/routes/Project/routes/Actions/components/RecentActions/NoRecentActions.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import Paper from 'material-ui/Paper'
+import classes from './RecentActions.scss'
+
+export const NoRecentActions = () => (
+  <Paper className={classes.empty}>
+    <span>No Recent Actions Found</span>
+  </Paper>
+)
+
+export default NoRecentActions
