+++ src/routes/Project/routes/Actions/components/RecentActions/RecentActions.enhancer.js
@@ -1,14 +1,16 @@
-import { get, orderBy, map } from 'lodash'
+import { get, map } from 'lodash'
 import { compose, withHandlers, withProps } from 'recompose'
 import { firestoreConnect, firebaseConnect } from 'react-redux-firebase'
 import { connect } from 'react-redux'
 import { initialize } from 'redux-form'
-import { spinnerWhileLoading } from 'utils/components'
+import { spinnerWhileLoading, renderWhileEmpty } from 'utils/components'
+import NoRecentActions from './NoRecentActions'
+import { databaseURLToProjectName } from 'utils'
 import { formNames } from 'constants'
 
 export default compose(
-  // Map redux state to props
   firebaseConnect(['displayNames']),
+  // Map redux state to props
   firestoreConnect(({ params, auth }) => [
     // Project environments
     {
@@ -16,6 +18,7 @@ export default compose(
       doc: params.projectId,
       subcollections: [{ collection: 'events' }],
       where: ['eventType', '==', 'requestActionRun'],
+      orderBy: ['createdAt', 'desc'],
       limit: 3,
       storeAs: 'recentActions'
     }
@@ -25,21 +28,27 @@ export default compose(
     displayNames: get(state.firebase, 'data.displayNames'),
     recentActions: get(state.firestore, `ordered.recentActions`)
   })),
+  renderWhileEmpty(['recentActions'], NoRecentActions),
   spinnerWhileLoading(['recentActions']),
   withProps(({ recentActions, displayNames }) => ({
-    orderedActions: map(
-      orderBy(recentActions, ['createdAt'], ['desc']),
-      event => {
-        const createdBy = get(event, 'createdBy')
-        if (createdBy) {
-          return {
-            ...event,
-            createdBy: get(displayNames, createdBy, createdBy)
-          }
+    orderedActions: map(recentActions, event => {
+      const createdBy = get(event, 'createdBy')
+      if (createdBy) {
+        return {
+          ...event,
+          createdBy: get(displayNames, createdBy, createdBy)
         }
-        return event
       }
-    )
+      return event
+    }),
+    actionToEnvironments: action => ({
+      src: databaseURLToProjectName(
+        get(action, 'eventData.inputValues.0.databaseURL', '')
+      ),
+      dest: databaseURLToProjectName(
+        get(action, 'eventData.inputValues.1.databaseURL', '')
+      )
+    })
   })),
   withHandlers({
     rerunAction: props => action => {
