+++ src/routes/Project/routes/Actions/components/RecentActions/RecentActions.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Paper from 'material-ui/Paper'
-import { map, startCase } from 'lodash'
+import { get, map, startCase } from 'lodash'
 import Table, {
   TableBody,
   TableCell,
@@ -14,7 +14,11 @@ import RedoIcon from 'material-ui-icons/Redo'
 import { formatDateTime } from 'utils/formatters'
 import classes from './RecentActions.scss'
 
-export const RecentActions = ({ orderedActions, rerunAction }) => (
+export const RecentActions = ({
+  orderedActions,
+  rerunAction,
+  actionToEnvironments
+}) => (
   <div className={classes.container}>
     <Paper>
       <Table className={classes.table}>
@@ -22,6 +26,9 @@ export const RecentActions = ({ orderedActions, rerunAction }) => (
           <TableRow>
             <TableCell>Time</TableCell>
             <TableCell>Created By</TableCell>
+            <TableCell>Template</TableCell>
+            <TableCell>Source/Dest</TableCell>
+            <TableCell>Path</TableCell>
             <TableCell>Redo</TableCell>
           </TableRow>
         </TableHead>
@@ -33,6 +40,19 @@ export const RecentActions = ({ orderedActions, rerunAction }) => (
                 <span>
                   {action.createdBy || startCase(action.createdByType)}
                 </span>
+              </TableCell>
+              <TableCell>
+                <span>{get(action, 'eventData.template.name', 'No Name')}</span>
+              </TableCell>
+              <TableCell>
+                <span>
+                  Source: {actionToEnvironments(action).src}
+                  <br />
+                  Dest: {actionToEnvironments(action).dest}
+                </span>
+              </TableCell>
+              <TableCell>
+                <span>{get(action, 'eventData.inputValues.2', 'No Path')}</span>
               </TableCell>
               <TableCell>
                 <Tooltip
@@ -61,6 +81,7 @@ export const RecentActions = ({ orderedActions, rerunAction }) => (
 
 RecentActions.propTypes = {
   orderedActions: PropTypes.array,
+  actionToEnvironments: PropTypes.func.isRequired,
   rerunAction: PropTypes.func.isRequired
 }
